refactor : assertThrow 메서드 입력을 Supplier에서 ThrowingCallable로 변경
