+++ src/main/java/assertv/assertions/ThrowingCallable.java
@@ -1,0 +1,5 @@
+package main.java.assertv.assertions;
+
+public interface ThrowingCallable {
+    void call() throws Throwable;
+}
